Move isRelativeGridTrackBreadthAsAuto off GridTrackSizingAlgorithm. https://bugs.webkit.org/show_bug.cgi?id=297245 rdar://problem/158074814
Reviewed by Brandon Stewart and Vitor Roriz.

This function currently lives on GridTrackSizingAlgorithm but there is
no need for that since it's used to determine some sort of
characteristic of a grid track's style. All that is required to make
this determination is the track itself and any available space. We can
instead move this over to GridLayoutFunctions and have it take in the
available space as an argument. This will make it a bit easier to move
some other functions off of GridTrackSizingAlgorithm which depend on
this one (calculateGridTrackSize and rawGridTrackSize). Since the logic
of those functions is a bit more involved, I don't want to include them
in this patch to keep the diffs for each patch focused.

Canonical link: https://commits.webkit.org/298554@main