[CoordinatedGraphics] Remove GLContext dependency on PlatformDisplay https://bugs.webkit.org/show_bug.cgi?id=297215
Reviewed by Miguel Gomez.

GLContext is keeping a plain reference to PlatformDisplay because it was
assumed that global shared display was the only possible
PlatformDisplay. This patch makes GLDisplay reference counted and
GLContext now takes a weak reference of a GLDisplay instead of a plain
reference of a PlatformDisplay. We still keep the constructors taking a
PlatformDisplay for convenience.

Canonical link: https://commits.webkit.org/298560@main