Remove Quirk ipadMini UA for open.spotify.com https://bugs.webkit.org/show_bug.cgi?id=297259 rdar://157503895
Reviewed by Brandon Stewart.

open.spotify.com had an issue on iPad where the player was not properly
working. In Bug 268928, a UA override was added to spotify to send the
iPad Mini UA. Right now, Spotify detects this UA and says that the
website is not supported on iPadMini UA.

Let's remove the UA override. The player is working with the desktop UA.
The new iPad window system also makes it easier to manipulate the bottom
scrubber for the player.

* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::needsIPadMiniUserAgent):

Canonical link: https://commits.webkit.org/298591@main