AX: Scrolled iframe content cannot be hit tested https://bugs.webkit.org/show_bug.cgi?id=297296 rdar://155985059
Reviewed by Tyler Wilcock.

When hit testing the main frame/page via WKAccessibilityWebPageObjectMac,
we account for the scroll position of that page. But, when an iframe is
hit tested, we are not adjusting for that frame's scroll position as
well. This fixes that issue, resolving several hit testing issues across ATs.

* LayoutTests/accessibility/mac/iframe-hit-testing-expected.txt: Added.
* LayoutTests/accessibility/mac/iframe-hit-testing.html: Added.
* LayoutTests/platform/mac-wk1/TestExpectations:
* Source/WebCore/accessibility/AccessibilityObject.cpp:
(WebCore::AccessibilityObject::elementAccessibilityHitTest const):

Canonical link: https://commits.webkit.org/298627@main