Fix flex-aspect-ratio-img-column-012 https://bugs.webkit.org/show_bug.cgi?id=297299 rdar://158173811
Reviewed by Alan Baradlay.

When computing the base size of a flex item, we need to make sure to
ignore any min/max main sizes to match the spec. When a replaced element
is a flex item, we currently handle this case correctly in the flex
item's inline direction but not the block direction. We can fix this by
doing the exact same thing that computeReplacedLogicalWidthRespectinMinMaxWidth
is doing by using a helper function, shouldIgnoreLogicalMinMaxHeightSizes,
which checks for this case.

Canonical link: https://commits.webkit.org/298629@main