[macOS Tahoe] fast/forms/range/input-appearance-range-rtl.html is an ImageOnlyFailure https://bugs.webkit.org/show_bug.cgi?id=297342 rdar://158241596
Reviewed by Abrar Rahman Protyasha.

Prior to macOS Tahoe, <input type=range> did not fill the track to match the
value. This made it possible to write a reference test matching LTR and RTL
layout. However, now that the track is filled, it is no longer possible to
write a reference match test.

As is already the case on iOS, where the fill has always been painted,
make the test a reference mismatch test.

* LayoutTests/fast/forms/range/input-appearance-range-rtl-expected-mismatch.html: Renamed from LayoutTests/platform/ios/fast/forms/range/input-appearance-range-rtl-expected-mismatch.html.
* LayoutTests/platform/mac-sequoia/TestExpectations:
* LayoutTests/platform/mac-sequoia/fast/forms/range/input-appearance-range-rtl-expected.html: Renamed from LayoutTests/fast/forms/range/input-appearance-range-rtl-expected.html.
* LayoutTests/platform/mac/TestExpectations:

Canonical link: https://commits.webkit.org/298648@main