Fix css-flexbox/flex-aspect-ratio-img-row-015 rdar://84995186 https://bugs.webkit.org/show_bug.cgi?id=232408
Reviewed by Alan Baradlay.

In the WPT, we have a row flexbox with a single flex item which is an SVG
which does not have an intrinsic size but has an intrinsic aspect ratio.
When computing the base size of this flex item, we compute its preferred
logical width.

Since this SVG does not have any intrinsic sizes, we end up computing the
stretch fit. When this occurs, we also need to consider any min/max
constraints imposed upon the item from the opposite axis. This ends up
constraining the logical width, and as a result, the base size, instead
of it being as wide as its containing block. We can achieve this by
using computeMinMaxLogicalWidthFromAspectRatio which computes these
transferred constraints and then clamping the stretch fit size between
these values.

Canonical link: https://commits.webkit.org/298651@main