Tabs loaded in the background sometimes show magenta flashes when switching to them https://bugs.webkit.org/show_bug.cgi?id=297344 rdar://157992869
Reviewed by Tim Horton.

When Command-clicking a link in Safari, the page is loaded in a background tab. This load does some initial
rendering and layer tree construction in a brief amount of time before the process is suspended. This first
rendering can do some drawing into an IOSurface, but as part of process suspension, we also make those
IOSurfaces volatile, and they can get purged under memory pressure.

When we switch back to the tab, we can currently parent the existing layer tree before we've had a chance
to make the surfaces non-volatile and render into them, and this can submit layers with purged IOSurfaces
to WindowServer. Previously, this resulted in a black flash. Now that IOSurface compression is enabled,
this is a magenta flash, which is much more obvious.

The fix is to call `hideContentUntilAnyUpdate()` on the drawing area if the web view is newly visible,
which disconnects the root remote layer tree node (thus unparenting its layer). This prevents us from
connecting a zombie layer tree until the next commit comes in, which is guaranteed to reconnect the
root layer.

Now that this is fixed, we re-enabled IOSurface compression on macOS.

* Source/WebKit/Shared/Cocoa/WebPreferencesDefaultValuesCocoa.mm:
(WebKit::defaultIOSurfaceLosslessCompressionEnabled):
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::dispatchActivityStateChange):

Canonical link: https://commits.webkit.org/298652@main