Implement Wasm js-string builtins https://bugs.webkit.org/show_bug.cgi?id=292630 rdar://150798623
Reviewed by Keith Miller.

This patch implements the Wasm js-string builtin proposal
(https://github.com/WebAssembly/js-string-builtins).

Availability of this feature is controlled by the `useWasmJSStringBuiltins` option,
`true` by default.

The relevant WPT test expectation files are changed to expect the tests to pass.

`FixedBitVector::test()` function is marked as `const` by this patch. It does
not require a mutable `this`, and the patch needs to use it on a vector
coming from a `const ModuleInformation&`.

Canonical link: https://commits.webkit.org/298653@main