[macOS Tahoe] fast/forms/range/slider-in-multi-column.html is a constant failure https://bugs.webkit.org/show_bug.cgi?id=297379 rdar://154997421
Reviewed by Richard Robinson and Abrar Rahman Protyasha.

The failing test uses a hardcoded offset to drag the slider thumb and then
asserts that the slider has a specific value.

However, this approach is flawed since the slider value is function of the
both the position and the width of thumb. The width of the thumb affects
the current value, since no part of the thumb can be dragged beyond the bounds
of the track.

The width of the slider thumb changed with the form control redesign in
macOS Tahoe, and the hardcoded offset no longer results in the same slider
value. Fix by computing the offset necessary to achieve the desired value.

* LayoutTests/fast/forms/range/slider-in-multi-column.html:

2 is subtracted from the expected value in order to continue to test the `step`.

* LayoutTests/platform/mac-sequoia/TestExpectations:
* LayoutTests/platform/mac/TestExpectations:

Canonical link: https://commits.webkit.org/298673@main