[CoordinatedGraphics] Prefer to use ANGLE API to create the fence for WebGL https://bugs.webkit.org/show_bug.cgi?id=297267
Reviewed by Alejandro G. Castro.

We currently use GLFence which is more convenient and works, but if
exportable fences are supported, we could use the ANGLE api to create
the fence, export it to and fd and then import it from the compositor.
Using ANGLE API we ensure all internal state is handled in sync when the
fence is added to the context.

Canonical link: https://commits.webkit.org/298675@main