[visionOS] Don't position the InteractionRegion container above UIKit-owned views https://bugs.webkit.org/show_bug.cgi?id=297331 <rdar://158220743>
Reviewed by Simon Fraser.

Some UIKit-owned views, like scroll indicators, are automatically
re-ordered to be on top of everything else.
When we update the position of InteractionRegion containers, we need to
make sure we're not causing extra re-ordering by constantly trying to
put the container on top of these views.

* Source/WebKit/UIProcess/RemoteLayerTree/RemoteLayerTreeNode.mm:
(WebKit::RemoteLayerTreeNode::repositionInteractionRegionsContainerIfNeeded):
When we encounter a view that doesn't have a corresponding node, break
instead of skipping.

This doesn't impact existing tests.

Canonical link: https://commits.webkit.org/298676@main