OrderedHashTable addImpl should only increment the alive entry count if no exception takes place https://bugs.webkit.org/show_bug.cgi?id=293579 rdar://152034907
Reviewed by Yusuke Suzuki.

If an exception is thrown while we are calculating the normalized key, we
will end up in a state where we have incremented aliveEntryCount, but the
map will not actually have an element added. We should only increment the
aliveEntryCount if we successfully add an element to the map.

* Source/JavaScriptCore/runtime/OrderedHashTableHelper.h:
(JSC::OrderedHashTableHelper::addImpl):

Originally-landed-as: 289651.538@safari-7621-branch (44dd2c945c05). rdar://157790926
Canonical link: https://commits.webkit.org/298709@main