4X fast/forms/switch/zoom-approximates-transform layout tests are constant ImageOnlyFailures https://bugs.webkit.org/show_bug.cgi?id=297426 rdar://154990654
Reviewed by Tim Horton.

Scale and zoom are unique operations which are not strictly required to look
identical to one another. Similar behavior exists for other controls besides
switch controls, and such behavior is not unique to WebKit. Add pixel tolerance
to each of the failing tests to account for this.

* LayoutTests/fast/forms/switch/zoom-approximates-transform-rtl.html:
* LayoutTests/fast/forms/switch/zoom-approximates-transform-vertical-lr.html:
* LayoutTests/fast/forms/switch/zoom-approximates-transform-vertical-rl.html:
* LayoutTests/fast/forms/switch/zoom-approximates-transform.html:
* LayoutTests/platform/mac-sequoia/TestExpectations:
* LayoutTests/platform/mac/TestExpectations:

Canonical link: https://commits.webkit.org/298714@main