iOS: Magenta flash over web content https://bugs.webkit.org/show_bug.cgi?id=297463 rdar://158419789
Reviewed by Tim Horton.

This change applies the fix made for macOS in 298652@main to iOS as well, since we now
have evidence that the same bug occurs there too; we can parent a layer tree that references
purged IOSurfaces, which, now with compression enabled, are magenta, not black.

The fix is just to remove the #ifdefs.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::dispatchActivityStateChange):

Canonical link: https://commits.webkit.org/298784@main