[iOS] Find on page highlight is broken for -webkit-user-select:none styled elements (affects many sites, including nytimes.com, bizjournal.com, cnn.com, developer.mozilla.org) https://bugs.webkit.org/show_bug.cgi?id=297441 rdar://8268929
Reviewed by Aditya Keerthi.

When we try and make a text indicator for a selection that was a
range that is webkit-user-select-none then the selection collapses
and the indicator does not have an associated image.

We should fallback to creating an indicator of the range painted
without the alterations for the indicator, for now. This will allow us
to at least have somewhat of an indicator, even if it is not as obvious.

We also shouldn't block on the existence of a selection rect for painting,
as if we have text rects and an image, that is also workable for showing
an indicator.

* Source/WebCore/page/TextIndicator.cpp:
(WebCore::TextIndicator::createWithRange):
* Source/WebKit/WebProcess/WebPage/WebFoundTextRangeController.cpp:
(WebKit::WebFoundTextRangeController::drawRect):

Canonical link: https://commits.webkit.org/298785@main