Add performJITMemcpyAtomic and simplify jit copying code [2]. https://bugs.webkit.org/show_bug.cgi?id=297235
Reviewed by Keith Miller.

The eventual goal of this series of patches is to make clear, explicit and separate
entrypoints into code that eventually copies to the jit region or an assembler buffer,
to clearly separate patching that must be atomic and patching that does not.

This is important to ensure that every place relying on atomic copying behaviour
is explicitly documented, so that ARMv7 can ensure the correct alignment of these regions.

To start, we carefuly thread through a new atomic flag, but we avoid changing behavior yet.

The only change this patch should make is that sometimes we can avoid a buffer allocation
when linking.

Canonical link: https://commits.webkit.org/298849@main