WebAssemblyCompileOptions::validateImportForBuiltinSetNames and WebAssemblyModuleRecord::initializeImports should check the import kind https://bugs.webkit.org/show_bug.cgi?id=297543 rdar://158575083
Reviewed by Yusuke Suzuki.

The function WebAssemblyCompileOptions::validateImportForBuiltinSetNames corresponds to
the spec algorithm 'To validate an import for builtins':
https://webassembly.github.io/js-string-builtins/js-api/#validate-an-import-for-builtins

The algorithm treats the import type as a self-contained entity held in 'imports[2]', and
in the last step enforces that it must be a function type by matching it against a
function type constructed in Step 6.

In our case the import type is fetched from one of the kind-specific arrays in moduleInfo,
so for equivalent behavior we must *first* check that the import kind is a function, and
then proceed with fetching its type and doing the type matching.

Canonical link: https://commits.webkit.org/298856@main