[import-w3c-tests] Support WPT working copy not called "web-platform-tests" https://bugs.webkit.org/show_bug.cgi?id=257730 rdar://problem/110278828
Reviewed by Brianna Fan.

The GitHub repo w3c/web-platform-tests was renamed to
web-platform-tests/wpt back in May 2018. Since then, `git clone <url>`
has created a directory called "wpt" from which we could not import.

This fix is, arguably, an inelegant solution to the problem: we simply
create a temporary directory, and create a symlink within it called
"web-platform-tests" pointing at the existing working copy.

This is done because we assume we're getting paths relative to
`LayoutTests/imported/w3c` all over the place and trying to
disentangle this has proven to be more work than its worth.

Longer term we should disentangle all of that, as part of furthering
the modern assumption that WPT is the only repo which we are importing
from, but fixing this usability bug is more urgent than that.

* Tools/Scripts/webkitpy/common/system/filesystem.py:
(FileSystem.symlink):
* Tools/Scripts/webkitpy/w3c/test_importer.py:
(parse_args):
(TestImporter.do_import):
(TestImporter._get_wpt_directory):
(TestImporter.should_skip_path):
These are all PurePath, because we (for now) require everything to go via the Filesystem object.
* Tools/Scripts/webkitpy/w3c/test_importer_unittest.py:
(TestImporterTest.test_import_dir_not_web_platform_tests):

Canonical link: https://commits.webkit.org/298861@main