Replace SerializedScriptValue roundtrip with an equivalent object filter https://bugs.webkit.org/show_bug.cgi?id=297564 rdar://158639867
Reviewed by Brady Eidson.

In order to have _WKJSHandle be able to be returned as a member of a dictionary or array,
we need to either implement serializing and deserializing of it through SerializedScriptValue
so it survives the roundtrip or remove the roundtrip and replace it with something functionally
equivalent.  I did the latter because it is also a performance improvement.
In order to maintain backwards compatibility, we need to keep all the weird quirks of
SerializedScriptValue when it deserializes into a non-JSDOMGlobalObject JSContext.
This is covered by many tests, such as TestWebKitAPI.EvaluateJavaScript.ReturnTypes,
TestWebKitAPI.WKWebView.EvaluateJavaScriptErrorCases, and many others.

* Source/WebCore/bindings/js/SerializedScriptValue.cpp:
(WebCore::SerializedScriptValue::deserializationBehavior):
* Source/WebCore/bindings/js/SerializedScriptValue.h:
* Source/WebKit/Shared/API/APISerializedScriptValue.cpp:
(API::SharedJSContextWK::singleton):
* Source/WebKit/Shared/JavaScriptEvaluationResult.cpp:
(WebKit::JavaScriptEvaluationResult::JSExtractor::addObjectToMap):
(WebKit::JavaScriptEvaluationResult::extract):
(WebKit::JavaScriptEvaluationResult::JSExtractor::toValue):
(WebKit::roundTripThroughSerializedScriptValue): Deleted.
* Source/WebKit/Shared/UserData.cpp:
* Source/WebKit/Sources.txt:
* Source/WebKit/SourcesGTK.txt:
* Source/WebKit/SourcesWPE.txt:
* Source/WebKit/UIProcess/API/APIInspectorExtension.cpp:
* Source/WebKit/UIProcess/API/C/WKPage.cpp:
* Source/WebKit/UIProcess/API/Cocoa/WKUserContentController.mm:
* Source/WebKit/UIProcess/API/Cocoa/WKWebView.mm:
* Source/WebKit/UIProcess/API/Cocoa/_WKInspectorExtension.mm:
* Source/WebKit/UIProcess/Extensions/Cocoa/API/WebExtensionContextAPIDevToolsInspectedWindow.mm:
* Source/WebKit/UIProcess/Inspector/WebInspectorUIExtensionControllerProxy.cpp:
* Source/WebKit/UIProcess/UserContent/WebUserContentControllerProxy.cpp:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:
* Source/WebKit/WebProcess/Extensions/API/Cocoa/WebExtensionAPIDevToolsInspectedWindowCocoa.mm:

Canonical link: https://commits.webkit.org/298898@main