REGRESSION(298653@main): ASSERTION FAILED: !m_needExceptionCheck https://bugs.webkit.org/show_bug.cgi?id=297556 rdar://158635946
Reviewed by Yusuke Suzuki.

A throw scope was added to function `webAssemblyInstantiateFunc` by the JS String builtins
work (https://bugs.webkit.org/show_bug.cgi?id=292630). Both branches of the final
`if-else` of the function can throw an exception. The original behavior of this
function was to leave the exception to be checked by the caller. To preserve this
behavior without triggering an exception check validation failure, the newly introduced
scope should be released before returning.

Canonical link: https://commits.webkit.org/298912@main