Use a more descriptive error message on an exception check validation failure https://bugs.webkit.org/show_bug.cgi?id=297604 rdar://158698040
Reviewed by Mark Lam.

An exception check validation failure in `VM::verifyExceptionCheckNeedIsSatisfied` is
currently triggered by `RELEASE_ASSERT(!m_needExceptionCheck)`. The assert condition is
invariably false because it appears in an `if` block guarded by `m_needExceptionCheck`.
Written this way it is not obvious that the failure is not a violation of internal VM
state consistency. It might even suggest that there used to be logic in this code,
now missing, that was expected to reset the flag to false.

The patch changes the assert to clarify that the unconditional failure is intentional
and explicitly state the reason.

Canonical link: https://commits.webkit.org/298929@main