[Tahoe] Switching tabs in Safari sometimes results in blank tabs https://bugs.webkit.org/show_bug.cgi?id=297630 rdar://158632267
Reviewed by Tim Horton.

When switching tabs via the mouse in Safari, there's a hover preview that uses a
_WKThumbnailView, and this persists until after the switch, when it's animated out. The
thumbnail view is associated with a WKWebView, and when present, can intercept the layer
passed to `WebViewImpl::setAcceleratedCompositingRootLayer()`, which prevents that layer
from being parented in the main view, resulting in blank contents.

298652@main changed the timing of the `setAcceleratedCompositingRootLayer()` call such
that it now happens while the thumbnail view is present, resulting in blank tabs until
Safari unparents the thumbnail view.

Fix this in the case where _WKThumbnailView has already generated the snapshot by having
`WebViewImpl::updateThumbnailViewLayer()` return a value indicating whether it actually
stole the root layer (i.e. when needing to generate the snapshot); if not, we parent the
root layer in the main view even when the _WKThumbnailView is present.

* Source/WebKit/UIProcess/mac/WebViewImpl.h:
* Source/WebKit/UIProcess/mac/WebViewImpl.mm:
(WebKit::WebViewImpl::setAcceleratedCompositingRootLayer):
(WebKit::WebViewImpl::updateThumbnailViewLayer):

Canonical link: https://commits.webkit.org/298935@main