Change StringBuilder in percentEncodeCharacters() to use RecordOverflow https://bugs.webkit.org/show_bug.cgi?id=297514 rdar://157533134
Reviewed by Alex Christensen.

Assigning a very large string to the username or password properties of
a URL object can result in a StringBuilder overflow during the call to
percentEncodeCharacters(), which will crash with the default overflow policy.
Change the policy from CrashOnOverflow to RecordOverflow.

* LayoutTests/fast/dom/DOMURL/url-very-large-password-no-crash-expected.txt: Added.
* LayoutTests/fast/dom/DOMURL/url-very-large-password-no-crash.html: Added.
* LayoutTests/fast/dom/DOMURL/url-very-large-username-no-crash-expected.txt: Added.
* LayoutTests/fast/dom/DOMURL/url-very-large-username-no-crash.html: Added.
* Source/WTF/wtf/URL.cpp:
(WTF::percentEncodeCharacters):

Canonical link: https://commits.webkit.org/298985@main