Fix TestWebKitAPI.WebKit.DidAssociateFormControls after 298736@main https://bugs.webkit.org/show_bug.cgi?id=297714 rdar://158830772
Unreviewed.

The test has an event when the form element is created, then another event
when the input element is added.  There was a race condition between the first event
and the navigation completion, which caused the test to sometimes check the wrong
alert, and it was not testing the event we care more about, which is when the
input element is added to the form.  Now it is!

This was pointed out to me by Arno Renevier.  Thanks, Arno!

* Tools/TestWebKitAPI/Tests/WebKitCocoa/FormValidation.mm:
(TestWebKitAPI::TEST(WebKit, DidAssociateFormControls)):

Canonical link: https://commits.webkit.org/298993@main