Modernize OrientationNotifier https://bugs.webkit.org/show_bug.cgi?id=297656 rdar://158765573
Reviewed by Jean-Yves Avenard.

We remove the need for OrientationNotifier to be a CheckedPtr.
We make OrientationNotifier use a weak hash set to handle observers instead of a Vector as it is simpler and safer to use.
To do so, we make OrientationNotifier::Observer a AbstractThreadSafeRefCountedAndCanMakeWeakPtr and we make RealtimeVideoCaptureSource a OrientationNotifier::Observer instead of its children classes.
A follow-up patch might relocate orientation handling in RealtimeVideoCaptureSource as well as a furuther improvement.

Canonical link: https://commits.webkit.org/299001@main